Update to skip to page 6 when requried.
If "never" is selected for all options in Question 2, ie: no substances were used in past 3 months then, Question 6 should be shown.
If only tobacco is NOT selected as "Never" in Q2, ie: if only tobacco is used in past 3 months then, skip to Question 6 after Question 4 because tobacco should not be displayed in Question 5.